# public/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDataDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponseDto"}}}}},"tags":["auth v1"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginLocalDataDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"tags":["auth v1"]}},"/api/v1/auth/logout":{"get":{"operationId":"AuthController_logout_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["auth v1"]}}},"info":{"title":" API - Documentation","description":"API for user interface development","version":"1.1","contact":{}},"tags":[],"servers":[{"url":"http://localhost:6005","description":"Local Development"},{"url":"https://47v5h1aj7h.execute-api.ap-southeast-1.amazonaws.com/dev","description":"DEV Environment (AWS Lambda)"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDataDto":{"type":"object","properties":{"email":{"type":"string"},"fullName":{"type":"string"},"password":{"type":"string"}},"required":["email","fullName","password"]},"RegisterResponseDto":{"type":"object","properties":{"data":{"type":"object","example":{"email":"example@gmail.com","other_fields":"..."},"nullable":true},"message":{"type":"string","example":"Success"},"statusCode":{"type":"number","example":201}},"required":["data","message","statusCode"]},"LoginLocalDataDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"data":{"type":"string","example":"JWT Token"},"message":{"type":"string","example":"Success"},"statusCode":{"type":"number","example":200}},"required":["data","message","statusCode"]}}}}